@@ -4,55 +4,54 @@ import { render, screen } from '@testing-library/react';
 import OrdersList from './ordersList';
 
 describe('Orders List', () => {
-    test('renders with no prop', () => {
-        render(
-            <OrdersList
-            />
-        )
-        const emptyDiv = document.getElementsByClassName('empty-orders');
-        const arr = Array.from(emptyDiv);
-        expect(arr.length).toBe(1);
-    });
+  test('renders with no prop', () => {
+    render(<OrdersList />);
+    const emptyDiv = document.getElementsByClassName('empty-orders');
+    const arr = Array.from(emptyDiv);
+    expect(arr.length).toBe(1);
+  });
 
-    test('renders one order', () => {
-        const orders = [
-            {
-                order_item: "Food",
-                quantity: "777",
-                _id: 1
-            }
-        ];
-        render(
-            <OrdersList
-                orders={orders}
-            />
-        )
-        expect(screen.getByText('Food')).toBeInTheDocument();
-        expect(screen.getByText(/^.*777.*$/gm)).toBeInTheDocument();
-    });
+  test('renders one order', () => {
+    const orders = [
+      {
+        order_item: 'Food',
+        quantity: '777',
+        _id: 1,
+      },
+    ];
+    render(<OrdersList orders={orders} />);
+    expect(screen.getByText('Food')).toBeInTheDocument();
+    expect(screen.getByText(/^.*777.*$/gm)).toBeInTheDocument();
+  });
 
-    test('renders multiple orders', () => {
-        const orders = [
-            {
-                order_item: "Food",
-                quantity: "777",
-                _id: 1
-            },
-            {
-                order_item: "Drink",
-                quantity: "888",
-                _id: 2
-            }
-        ];
-        render(
-            <OrdersList
-                orders={orders}
-            />
-        )
-        expect(screen.getByText('Food')).toBeInTheDocument();
-        expect(screen.getByText(/^.*777.*$/gm)).toBeInTheDocument();
-        expect(screen.getByText('Drink')).toBeInTheDocument();
-        expect(screen.getByText(/^.*888.*$/gm)).toBeInTheDocument();
+  test('renders multiple orders', () => {
+    const orders = [
+      {
+        order_item: 'Food',
+        quantity: '777',
+        _id: 1,
+      },
+      {
+        order_item: 'Drink',
+        quantity: '888',
+        _id: 2,
+      },
+    ];
+    render(<OrdersList orders={orders} />);
+    expect(screen.getByText('Food')).toBeInTheDocument();
+    expect(screen.getByText(/^.*777.*$/gm)).toBeInTheDocument();
+    expect(screen.getByText('Drink')).toBeInTheDocument();
+    expect(screen.getByText(/^.*888.*$/gm)).toBeInTheDocument();
+  });
 
-    });
-})
+  test('correct time display', () => {
+    const orders = [
+      {
+        createdAt: '2021-10-18T21:53:40.293Z',
+        _id: 1,
+      },
+    ];
+    render(<OrdersList orders={orders} />);
+    expect(screen.getByText(/^.*14:53:40.*$/gm)).toBeInTheDocument();
+  });
+});
