@@ -1,7 +1,11 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, compose } from 'redux';
 import ReduxThunk from 'redux-thunk';
 import reducers from './reducers';
-
-const store = createStore(reducers, {}, applyMiddleware(ReduxThunk));
-
+const store = createStore(
+  reducers,
+  compose(
+    applyMiddleware(ReduxThunk),
+    window.devToolsExtension ? window.devToolsExtension() : f => f
+  )
+);
 export default store;
