@@ -1,16 +1,17 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { Main, Login, OrderFormHook, ViewOrdersHook } from '../components';
+import { GuardedRoute } from '../components/login/GuardedRoute'
 
-const AppRouter = (props) => {
+const AppRouter = props => {
   return (
     <Router>
-      <Route path="/" exact component={Main} />
-      <Route path="/login" exact component={Login} />
-      <Route path="/order" exact component={OrderFormHook} />
-      <Route path="/view-orders" exact component={ViewOrdersHook} />
+      <Route path='/' exact component={Main} />
+      <Route path='/login' exact component={Login} />
+      <GuardedRoute path='/order' exact component={OrderFormHook} />
+      <GuardedRoute path='/view-orders' exact component={ViewOrdersHook} />
     </Router>
   );
-}
+};
 
 export default AppRouter;
